LayoutTests/http/tests/cookies/accept-partitioned-first-and-third-party-cookies.https-expected.txt

Tests that first and third-party partitioned cookies are accepted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Cookies are: test_cookie = 1

--------
Frame: '<!--frame2-->'
--------
Cookies are:

--------
Frame: '<!--frame3-->'
--------
Cookies are: test_cookie = 1

--------
Frame: '<!--frame4-->'
--------
Cookies are: test_cookie = 1

--------
Frame: '<!--frame5-->'
--------
